Improves the piece selection using a hit area.
Picking pieces by their blocks is fiddly on touch screens. The
original offsets the piece by 3 blocks so the hand doesn't obscure the
piece while dragging. That's ugly and breaks the interaction model.

We leave some room at the bottom/right so you can drag without having
the finger over the piece. It's like placing the piece on a pizza
shovel and moving it around.